# Under high load, defer flushing watches to reduce time spent in syscalls 