--- a/pkg/util/wait/delayed_action_test.go
+++ b/pkg/util/wait/delayed_action_test.go
@@ -17,6 +17,7 @@ limitations under the License.
 package wait
 
 import (
+	"math/rand"
 	"testing"
 	"time"
 )
@@ -84,4 +85,63 @@ func TestDelayedAction(t *testing.T) {
 	}
 	<-a.After()
 	a.Done()
+}
+
+func TestDelayedActionWait(t *testing.T) {
+	// delay when rate is over 10/ms
+	a := NewChannelDelayedAction(Interval{Count: 10, Interval: time.Millisecond}, Interval{Count: 100, Interval: 5 * time.Millisecond})
+
+	expectTries := 9
+
+	for i := 0; i < 500; i++ {
+		// kick into after
+		tries := 0
+		for {
+			if !a.Run() {
+				break
+			}
+			tries++
+		}
+		if tries != expectTries {
+			t.Fatalf("unexpected number of tries %d %d for %#v", expectTries, tries, a)
+		}
+
+		var times int
+		switch rand.Int31n(5) {
+		case 0:
+			times = 1
+		case 1:
+			times = 5 + int(rand.Int31n(10))
+		case 2:
+			times = 95 + int(rand.Int31n(10))
+		case 3:
+			times = 150 + int(rand.Int31n(1000))
+		default:
+			t.Logf("sleeping until after")
+			time.Sleep(5 + time.Duration(rand.Int31n(10))*time.Millisecond)
+
+			expectTries = 9
+		}
+
+		if times > 0 {
+			t.Logf("running %d times", times)
+			for j := 0; j < times; j++ {
+				if a.Run() {
+					t.Fatalf("should not switch back to defer until after Done(): %#v", a)
+				}
+			}
+			expectTries = 9 - times
+			if expectTries < 0 {
+				expectTries = 0
+			}
+		}
+
+		select {
+		case <-a.After():
+			a.Done()
+		case <-time.After(10 * time.Second):
+			// this could be flaky, we could set this even higher
+			t.Fatalf("timer did not fire within 1s: %#v", a)
+		}
+	}
 }