Has a significant effect on throughput for overloaded watches. In single
threaded tests watch spends 25% less overall time (33% more throughput)
due to less syscalls to flush. Should not affect overall latency, although
could stand to be tuned / discussed more.

@wojtek-t

Fixes #24729