@hongchaodeng 2016-04-25
Can you write some benchmarks?