@xiang90 2016-04-25
@smarterclayton Have you tried something simpler: like http://mechanical-sympathy.blogspot.com/2011/10/smart-batching.html? Basically we only do flush when there is nothing in the chan. We can do len(ch) to know the current number of items in the chan buffer.