@smarterclayton 2016-04-25
We already have benchmarks for watch:

```
$ go test ./pkg/apiserver -run=XXX -bench=BenchmarkWatchProtobuf -benchmem -test.count=1 -test.v -test.benchtime 15s
PASS
BenchmarkWatchProtobuf-8	  500000	     40995 ns/op	    8449 B/op	     106 allocs/op
ok  	k8s.io/kubernetes/pkg/apiserver	20.946s
$ git checkout queue_flush
$ go test ./pkg/apiserver -run=XXX -bench=BenchmarkWatchProtobuf -benchmem -test.count=1 -test.v -test.benchtime 15s
PASS
BenchmarkWatchProtobuf-8	 1000000	     31632 ns/op	    8443 B/op	     106 allocs/op
ok  	k8s.io/kubernetes/pkg/apiserver	31.987s
$ benchcmp /tmp/old /tmp/new
benchmark                    old ns/op     new ns/op     delta
BenchmarkWatchProtobuf-8     40995         31632         -22.84%

benchmark                    old allocs     new allocs     delta
BenchmarkWatchProtobuf-8     106            106            +0.00%

benchmark                    old bytes     new bytes     delta
BenchmarkWatchProtobuf-8     8449          8443          -0.07%
```