@smarterclayton 2016-04-25
Will look - time based flush amortizes over longer windows but is simpler, I agree.