@xiang90 2016-04-25
@smarterclayton Time based is good at reducing number of syscalls even if it is under light load. And it is probably also good at handling very heavy load. We will still batch based on policy anyway. 

The feedback based is good at reducing syscalls at some degree and also keeping good latency.