@smarterclayton 2016-04-25
Tried with len(watch) and it was never non-zero - I suspect we don't buffer that particular channel.