@xiang90 2016-04-25
> Tried with len(watch) and it was never non-zero - I suspect we don't buffer that particular channel. Will look into it later.

Probably there this is where we should improve. Or it blocks the entire watch pipeline directly from etcd/client. It might lead to missing event if it blocks long time.