@smarterclayton 2016-04-25
API watches are served by the etcd watch cache, so it's at least decoupled from etcd.  Agree we should have a buffer depth there.