@xiang90 2016-04-26
@smarterclayton 

I am not sure. So here is the code in etcd helper: https://github.com/kubernetes/kubernetes/blob/master/pkg/storage/etcd/etcd_watcher.go#L141

The outgoing chan has zero buffer. 

When etcd/client receives a update, it will just try to send stuff to that chan and it blocks: https://github.com/kubernetes/kubernetes/blob/master/pkg/storage/etcd/etcd_watcher.go#L362.

Do you mean that there is another in-mem consumer that actually buffer the events? So it does not directly blocks on the io stream to the client? Can you please point me where is that buffered consumer? I would like to give it a look too. Thanks!