@smarterclayton 2016-04-26
https://github.com/kubernetes/kubernetes/blob/master/pkg/storage/watch_cache.go