@wojtek-t 2016-04-26
@xiang90 Yes - this was the biggest optimization that allowed us to scale from 100 to 1000 bascially