@lavalamp 2016-04-26
Unit tests & godoc comments please :) otherwise lgtm