@smarterclayton 2016-04-26
```
2016/04/26 12:42:13 http: panic serving 127.0.0.1:53889: runtime error: invalid memory address or nil pointer dereference
goroutine 4127 [running]:
net/http.(*conn).serve.func1(0xc82901c280)
	/usr/local/go/src/net/http/server.go:1389 +0xc1
panic(0x1d41ec0, 0xc8200100c0)
	/usr/local/go/src/runtime/panic.go:426 +0x4e9
k8s.io/kubernetes/test/integration.TestQuota.func1(0x7fc0e9dc5158, 0xc826b64f70, 0xc82355d180)
	/go/src/k8s.io/kubernetes/_output/local/go/src/k8s.io/kubernetes/test/integration/quota_test.go:58 +0x1a
net/http.HandlerFunc.ServeHTTP(0xc828ff1610, 0x7fc0e9dc5158, 0xc826b64f70, 0xc82355d180)
	/usr/local/go/src/net/http/server.go:1618 +0x3a
net/http.serverHandler.ServeHTTP(0xc828fbe200, 0x7fc0e9dc5158, 0xc826b64f70, 0xc82355d180)
	/usr/local/go/src/net/http/server.go:2081 +0x19e
net/http.(*conn).serve(0xc82901c280)
	/usr/local/go/src/net/http/server.go:1472 +0xf2e
created by net/http.(*Server).Serve
```

is not me