@smarterclayton 2016-04-30
All comments addressed, added even more tests.