@wojtek-t 2016-05-04
I'm not convinved about the values, but we can probably tune them later.

LGTM