@smarterclayton 2016-05-04
@k8s-bot test this issue #25139

Not sure if this is related, will test a few more times.