@smarterclayton 2016-05-04
Removing LGTM till I'm sure this is a not me flake