@smarterclayton 2016-05-04
Have not seen flake recur.