@smarterclayton 2016-05-05
```
STEP: Creating a pod to test override command
May  4 20:59:15.138: INFO: Failed to create pod: pods "client-containers-bbb62802-1275-11e6-aaae-42010af0000d" is forbidden: no API token found for service account e2e-tests-containers-qttnm/default, retry after the token is automatically created and added to the service account
```

@liggitt will need to reflect with you on how slightly delayed watches (flushes happening less often) could result in this race.