@liggitt 2016-05-05
> @liggitt will need to reflect with you on how slightly delayed watches (flushes happening less often) could result in this race.

e2e namespace setup waits until it sees a service account and token (via polling lists/gets, I think). The actual list-store that is used by admission is watch-driven, so delaying watch flushes could make the admission plugin cache lag