@smarterclayton 2016-05-05
In this case during e2e we might have bursts up to 10/sec, which might be enough to kick a user into the 10ms delay for flush regime, but it seems surprising we would be that different unless there's something in the controllers themselves missing an update chance.