@liggitt 2016-05-05
haven't reviewed the change closely, but you're sure this isn't introducing a long watch delay accidentally?