@smarterclayton 2016-05-05
Fairly sure - it would show up in other ways if so (I had that bug earlier
and 100 tests failed)

On Thu, May 5, 2016 at 1:02 AM, Jordan Liggitt <notifications@github.com>
wrote:

> haven't reviewed the change closely, but you're sure this isn't
> introducing a long watch delay accidentally?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly or view it on GitHub
> <https://github.com/kubernetes/kubernetes/pull/24768#issuecomment-217077061>
>