@smarterclayton 2016-05-05
Experimenting with the API token admission forcing the client to retry (because data isn't available yet)