@smarterclayton 2016-05-05
@k8s-bot test this issue #IGNORE