@smarterclayton 2016-05-06
@k8s-bot test this issue #IGNORE