@smarterclayton 2016-05-09
@k8s-bot test this issue #25330 (might be unrelated flake)