@wojtek-t 2016-05-09
@smarterclayton - what's the current status of it? I mean, there seemed to be a problem with this PR. Do you think it's fixed now?