@smarterclayton 2016-05-09
Current status is I believe I've fixed the race condition in the e2e test, but since it was ~0.2-0.4 probability I needed a lot of runs.  Going to keep running today - @liggitt can you review the changes I made to your controller, and I'll add a test to verify the new retry-after behavior from the server and that the client respects it.