@smarterclayton 2016-05-09
@k8s-bot test this issue #IGNORE (trying to recreate the flake)